+++ Runtime/Components/VStackComponent.cs
@@ -1,0 +1,25 @@
+using UnityEngine;
+using UnityEngine.UI;
+
+namespace UCGUI
+{
+    public class VStackComponent : HorizontalOrVerticalLayoutComponent
+    {
+        public override void Awake()
+        {
+            base.Awake();
+            AddVerticalLayout();
+        }
+
+        protected override HorizontalOrVerticalLayoutGroup GetLayout()
+        {
+            return VerticalLayout;
+        }
+
+        public override void Start()
+        {
+            base.Start();
+            this.DisplayName("VStack");
+        }
+    }
+}